+++ tests/tine20/Tinebase/Export/XlsxTest.php
@@ -341,6 +341,12 @@ class Tinebase_Export_XlsxTest extends TestCase
 
     public function testConvertToPdf()
     {
+        if (Tinebase_Config::getInstance()->{Tinebase_Config::FILESYSTEM}->{Tinebase_Config::FILESYSTEM_CREATE_PREVIEWS} != true
+            || Tinebase_Config::getInstance()->{Tinebase_Config::FILESYSTEM}->{Tinebase_Config::FILESYSTEM_PREVIEW_SERVICE_VERSION}  < 2
+        ) {
+            $this->markTestSkipped('no docservice configured');
+        }
+
         /** @var Addressbook_Export_Xls $export */
         $export = Tinebase_Export::factory(new Addressbook_Model_ContactFilter(),
             [
@@ -356,13 +362,15 @@ class Tinebase_Export_XlsxTest extends TestCase
 
         $export->generate();
 
+        $file = null;
         try {
             $file = $export->convert(Tinebase_Export_Convertible::PDF);
-
             $this->assertEquals('application/pdf', mime_content_type($file));
 
         } finally {
-            unlink($file);
+            if ($file) {
+                unlink($file);
+            }
         }
     }
 }
